implement parallelized proof reading
