+++ src/groth16/mod.rs
@@ -2,12 +2,6 @@
 //!
 //! [Groth16]: https://eprint.iacr.org/2016/260
 
-use crate::bls::Engine;
-
-use groupy::{CurveAffine, EncodedPoint};
-
-use std::io::{self, Read, Write};
-
 #[cfg(test)]
 mod tests;
 
@@ -15,182 +9,17 @@ mod ext;
 mod generator;
 mod mapped_params;
 mod params;
+mod proof;
 mod prover;
+mod utils;
 mod verifier;
 mod verifying_key;
-
-mod utils;
 
 pub use self::ext::*;
 pub use self::generator::*;
 pub use self::mapped_params::*;
+pub use self::params::*;
+pub use self::proof::*;
 pub use self::prover::*;
 pub use self::verifier::*;
 pub use self::verifying_key::*;
-pub use params::*;
-
-#[derive(Clone, Debug)]
-pub struct Proof<E: Engine> {
-    pub a: E::G1Affine,
-    pub b: E::G2Affine,
-    pub c: E::G1Affine,
-}
-
-impl<E: Engine> PartialEq for Proof<E> {
-    fn eq(&self, other: &Self) -> bool {
-        self.a == other.a && self.b == other.b && self.c == other.c
-    }
-}
-
-impl<E: Engine> Proof<E> {
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        writer.write_all(self.a.into_compressed().as_ref())?;
-        writer.write_all(self.b.into_compressed().as_ref())?;
-        writer.write_all(self.c.into_compressed().as_ref())?;
-
-        Ok(())
-    }
-
-    pub fn read<R: Read>(mut reader: R) -> io::Result<Self> {
-        let mut g1_repr = <E::G1Affine as CurveAffine>::Compressed::empty();
-        let mut g2_repr = <E::G2Affine as CurveAffine>::Compressed::empty();
-
-        reader.read_exact(g1_repr.as_mut())?;
-        let a = g1_repr
-            .into_affine()
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-            .and_then(|e| {
-                if e.is_zero() {
-                    Err(io::Error::new(
-                        io::ErrorKind::InvalidData,
-                        "point at infinity",
-                    ))
-                } else {
-                    Ok(e)
-                }
-            })?;
-
-        reader.read_exact(g2_repr.as_mut())?;
-        let b = g2_repr
-            .into_affine()
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-            .and_then(|e| {
-                if e.is_zero() {
-                    Err(io::Error::new(
-                        io::ErrorKind::InvalidData,
-                        "point at infinity",
-                    ))
-                } else {
-                    Ok(e)
-                }
-            })?;
-
-        reader.read_exact(g1_repr.as_mut())?;
-        let c = g1_repr
-            .into_affine()
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-            .and_then(|e| {
-                if e.is_zero() {
-                    Err(io::Error::new(
-                        io::ErrorKind::InvalidData,
-                        "point at infinity",
-                    ))
-                } else {
-                    Ok(e)
-                }
-            })?;
-
-        Ok(Proof { a, b, c })
-    }
-}
-
-#[cfg(test)]
-mod test_with_bls12_381 {
-    use super::*;
-    use crate::{Circuit, ConstraintSystem, SynthesisError};
-
-    use crate::bls::{Bls12, Fr};
-    use ff::Field;
-    use rand::thread_rng;
-
-    #[test]
-    fn serialization() {
-        struct MySillyCircuit<E: Engine> {
-            a: Option<E::Fr>,
-            b: Option<E::Fr>,
-        }
-
-        impl<E: Engine> Circuit<E> for MySillyCircuit<E> {
-            fn synthesize<CS: ConstraintSystem<E>>(
-                self,
-                cs: &mut CS,
-            ) -> Result<(), SynthesisError> {
-                let a = cs.alloc(|| "a", || self.a.ok_or(SynthesisError::AssignmentMissing))?;
-                let b = cs.alloc(|| "b", || self.b.ok_or(SynthesisError::AssignmentMissing))?;
-                let c = cs.alloc_input(
-                    || "c",
-                    || {
-                        let mut a = self.a.ok_or(SynthesisError::AssignmentMissing)?;
-                        let b = self.b.ok_or(SynthesisError::AssignmentMissing)?;
-
-                        a.mul_assign(&b);
-                        Ok(a)
-                    },
-                )?;
-
-                cs.enforce(|| "a*b=c", |lc| lc + a, |lc| lc + b, |lc| lc + c);
-
-                Ok(())
-            }
-        }
-
-        let rng = &mut thread_rng();
-
-        let params =
-            generate_random_parameters::<Bls12, _, _>(MySillyCircuit { a: None, b: None }, rng)
-                .unwrap();
-
-        {
-            let mut v = vec![];
-
-            params.write(&mut v).unwrap();
-            assert_eq!(v.len(), 2136);
-
-            let de_params = Parameters::read(&v[..], true).unwrap();
-            assert!(params == de_params);
-
-            let de_params = Parameters::read(&v[..], false).unwrap();
-            assert!(params == de_params);
-        }
-
-        let pvk = prepare_verifying_key::<Bls12>(&params.vk);
-
-        for _ in 0..100 {
-            let a = Fr::random(rng);
-            let b = Fr::random(rng);
-            let mut c = a;
-            c.mul_assign(&b);
-
-            let proof = create_random_proof(
-                MySillyCircuit {
-                    a: Some(a),
-                    b: Some(b),
-                },
-                &params,
-                rng,
-            )
-            .unwrap();
-
-            let mut v = vec![];
-            proof.write(&mut v).unwrap();
-
-            assert_eq!(v.len(), 192);
-
-            let de_proof = Proof::read(&v[..]).unwrap();
-            assert!(proof == de_proof);
-
-            assert!(verify_proof(&pvk, &proof, &[c]).unwrap());
-            assert!(!verify_proof(&pvk, &proof, &[a]).unwrap());
-        }
-    }
-}
